Renamed project etc to ReactNativeDeferredTemplate
