+++ ReactNativeDeferredTemplate/android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-	package="com.salesforce.reactnativetemplate"
+	package="com.salesforce.reactnativedeferredtemplate"
 	android:versionCode="1"
 	android:versionName="1.0"
 	android:installLocation="internalOnly">
@@ -65,9 +65,9 @@
             The URI constructed from this must be unique and will be set in the SP app
             as well. This will be used to launch the IDP app from the SP app.
 
-            In the example given below, 'scheme' is "com.salesforce.reactnativetemplate" and
+            In the example given below, 'scheme' is "com.salesforce.reactnativedeferredtemplate" and
             'host' is "authenticator". The full URI that needs to be set on the SP app would be
-            "com.salesforce.reactnativetemplate://authenticator".
+            "com.salesforce.reactnativedeferredtemplate://authenticator".
         -->
 		<!--
         <activity android:name="com.salesforce.androidsdk.auth.idp.IDPAccountPickerActivity"
@@ -76,7 +76,7 @@
 
 
             <intent-filter>
-                <data android:scheme="com.salesforce.reactnativetemplate"
+                <data android:scheme="com.salesforce.reactnativedeferredtemplate"
                     android:host="authenticator" />
                 <action android:name="android.intent.action.VIEW" />
                 <category android:name="android.intent.category.BROWSABLE" />
@@ -93,9 +93,9 @@
         receive the messages and registration result. This must be of
         the format <app-package-name>.C2D_MESSAGE for it to work.
     -->
-    <permission android:name="com.salesforce.reactnativetemplate.C2D_MESSAGE"
+    <permission android:name="com.salesforce.reactnativedeferredtemplate.C2D_MESSAGE"
         android:protectionLevel="signature" />
 
-    <uses-permission android:name="com.salesforce.reactnativetemplate.C2D_MESSAGE" />
+    <uses-permission android:name="com.salesforce.reactnativedeferredtemplate.C2D_MESSAGE" />
     <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW"/>
 </manifest>
