+++ ReactNativeDeferredTemplate/android/app/src/main/java/com/salesforce/reactnativedeferredtemplate/MainApplication.java
@@ -24,7 +24,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.salesforce.reactnativetemplate;
+package com.salesforce.reactnativedeferredtemplate;
 
 import android.app.Application;
