+++ ReactNativeDeferredTemplate/android/app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="account_type">com.salesforce.reactnativetemplate.login</string>
-	<string name="app_name">ReactNativeTemplate</string>
-	<string name="app_package">com.salesforce.reactnativetemplate</string>
+    <string name="account_type">com.salesforce.reactnativedeferredtemplate.login</string>
+	<string name="app_name">ReactNativeDeferredTemplate</string>
+	<string name="app_package">com.salesforce.reactnativedeferredtemplate</string>
 </resources>
