+++ ReactNativeDeferredTemplate/android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'ReactNativeTemplate'
+rootProject.name = 'ReactNativeDeferredTemplate'
 include ':react-native-safe-area-context'
 project(':react-native-safe-area-context').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-safe-area-context/android')
 include ':react-native-gesture-handler'
