+++ ReactNativeDeferredTemplate/ios/ReactNativeDeferredTemplate.xcodeproj/xcshareddata/xcschemes/ReactNativeDeferredTemplate.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "ReactNativeTemplate.app"
-               BlueprintName = "ReactNativeTemplate"
-               ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+               BuildableName = "ReactNativeDeferredTemplate.app"
+               BlueprintName = "ReactNativeDeferredTemplate"
+               ReferencedContainer = "container:ReactNativeDeferredTemplate.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ReactNativeTemplateTests.xctest"
-               BlueprintName = "ReactNativeTemplateTests"
-               ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+               BuildableName = "ReactNativeDeferredTemplateTests.xctest"
+               BlueprintName = "ReactNativeDeferredTemplateTests"
+               ReferencedContainer = "container:ReactNativeDeferredTemplate.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -45,9 +45,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ReactNativeTemplate.app"
-            BlueprintName = "ReactNativeTemplate"
-            ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+            BuildableName = "ReactNativeDeferredTemplate.app"
+            BlueprintName = "ReactNativeDeferredTemplate"
+            ReferencedContainer = "container:ReactNativeDeferredTemplate.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <Testables>
@@ -56,9 +56,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ReactNativeTemplateTests.xctest"
-               BlueprintName = "ReactNativeTemplateTests"
-               ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+               BuildableName = "ReactNativeDeferredTemplateTests.xctest"
+               BlueprintName = "ReactNativeDeferredTemplateTests"
+               ReferencedContainer = "container:ReactNativeDeferredTemplate.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -78,9 +78,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ReactNativeTemplate.app"
-            BlueprintName = "ReactNativeTemplate"
-            ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+            BuildableName = "ReactNativeDeferredTemplate.app"
+            BlueprintName = "ReactNativeDeferredTemplate"
+            ReferencedContainer = "container:ReactNativeDeferredTemplate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -95,9 +95,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ReactNativeTemplate.app"
-            BlueprintName = "ReactNativeTemplate"
-            ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+            BuildableName = "ReactNativeDeferredTemplate.app"
+            BlueprintName = "ReactNativeDeferredTemplate"
+            ReferencedContainer = "container:ReactNativeDeferredTemplate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
