+++ ReactNativeDeferredTemplate/ios/ReactNativeDeferredTemplate/AppDelegate.m
@@ -152,7 +152,7 @@
     jsCodeLocation = [[RCTBundleURLProvider sharedSettings] jsBundleURLForBundleRoot:@"index" fallbackResource:nil];
 
     RCTRootView *rootView = [[RCTRootView alloc] initWithBundleURL:jsCodeLocation
-                                                        moduleName:@"ReactNativeTemplate"
+                                                        moduleName:@"ReactNativeDeferredTemplate"
                                                  initialProperties:nil
                                                      launchOptions:self.launchOptions];
     rootView.backgroundColor = [[UIColor alloc] initWithRed:1.0f green:1.0f blue:1.0f alpha:1];
