+++ ReactNativeDeferredTemplate/ios/ReactNativeDeferredTemplate/ReactNativeDeferredTemplate.entitlements
@@ -1,10 +1,10 @@
-<?xml version="1.0" encoding="UTF-8"?>
+re<?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
 	<key>keychain-access-groups</key>
 	<array>
-		<string>$(AppIdentifierPrefix)com.salesforce.iosreactnativetemplate.ReactNativeTemplate</string>
+		<string>$(AppIdentifierPrefix)com.salesforce.iosreactnativedeferredtemplate.ReactNativeDeferredTemplate</string>
 	</array>
 </dict>
 </plist>
