+++ ReactNativeDeferredTemplate/ios/ReactNativeDeferredTemplate/main.m
