+++ ReactNativeDeferredTemplate/template.js
@@ -42,9 +42,9 @@ function prepare(config, replaceInFiles, moveFile, removeFile) {
         var path = require('path');
 
         // Values in template
-        var templateAppName = 'ReactNativeTemplate';
-        var templatePackageName = 'com.salesforce.reactnativetemplate';
-        var templateOrganization = 'ReactNativeTemplateOrganizationName';
+        var templateAppName = 'ReactNativeDeferredTemplate';
+        var templatePackageName = 'com.salesforce.reactnativedeferredtemplate';
+        var templateOrganization = 'ReactNativeDeferredTemplateOrganizationName';
 
         // Key files
         var templatePackageJsonFile = 'package.json';
@@ -101,8 +101,8 @@ function prepare(config, replaceInFiles, moveFile, removeFile) {
         var path = require('path');
 
         // Values in template
-        var templateAppName = 'ReactNativeTemplate';
-        var templatePackageName = 'com.salesforce.reactnativetemplate';
+        var templateAppName = 'ReactNativeDeferredTemplate';
+        var templatePackageName = 'com.salesforce.reactnativedeferredtemplate';
 
         // Key files
         var templatePackageJsonFile = 'package.json';
@@ -113,8 +113,8 @@ function prepare(config, replaceInFiles, moveFile, removeFile) {
         var templateAppBuildGradleFile = path.join('android', 'app', 'build.gradle');
         var templateStringsXmlFile = path.join('android', 'app', 'src', 'main', 'res', 'values', 'strings.xml');
         var templateBootconfigFile = path.join('android', 'app', 'src', 'main', 'res', 'values', 'bootconfig.xml');
-        var templateMainActivityFile = path.join('android', 'app', 'src', 'main', 'java', 'com', 'salesforce', 'reactnativetemplate', 'MainActivity.java');
-        var templateMainApplicationFile = path.join('android', 'app', 'src', 'main', 'java', 'com', 'salesforce', 'reactnativetemplate', 'MainApplication.java');
+        var templateMainActivityFile = path.join('android', 'app', 'src', 'main', 'java', 'com', 'salesforce', 'reactnativedeferredtemplate', 'MainActivity.java');
+        var templateMainApplicationFile = path.join('android', 'app', 'src', 'main', 'java', 'com', 'salesforce', 'reactnativedeferredtemplate', 'MainApplication.java');
 
         //
         // Replace in files
